requirements.txt
python-telegram-bot==20.7
aiogram==3.22.0
Pillow==10.0.0
redis==4.5.4